서버와 클라이언트 영역을 분리, server와 client로 나눔, server에는 서버에 대한 요청과 서버에서 구현할 비지니스 로직 부분 , [리퀘스트 요청 분리하여 controller, 서버의 비지니스 로직을 구현할 service] client에는 resources 를 관리, html 페이지 부분,
