+++ devleesk/client/resources/js/bootstrap/bootstrap.js
