+++ devleesk/client/views/layout.jade
@@ -2,8 +2,8 @@ doctype html
 html
 	head
 		title= title
-		link(rel='stylesheet', href='/stylesheets/style.css')
-		link(rel='stylesheet', href='/stylesheets/bootstrap/bootstrap.css')
+		link(rel='stylesheet', href='/css/style.css')
+		link(rel='stylesheet', href='/css/bootstrap/bootstrap.css')
 		script(src="/js/jquery/jquery-3.2.1.min.js")
 		script(src="/js/bootstrap/bootstrap.js")
 	body
