+++ devleesk/server/controller/chatting.js
@@ -3,3 +3,4 @@
 exports.chatting = function(req, res){
 	res.render('chatting', { title: 'chatting' });
 };
+
